// web/src/api/wire.ts
export type ExerciseType =
  | "Bodyweight"
  | "WeightedBodyweight"
  | "AssistedBodyweight"
  | "Weighted";
// ...
export type FullMicrocycle = {
  id: number;
  position: number;
  phase: Phase | null;
};

export type FullPlannedExercise = {
  id: number;
  exercise: LibraryExercise;
  position: number;
  prescriptions: Array<Prescription>;
};

export type FullWorkout = {
  id: number;
  name: string;
  position: number;
  planned_exercises: Array<FullPlannedExercise>;
};
// ...
export type LibraryExercise = {
  id: number;
  name: Name;
  exercise_type: ExerciseType;
  primary_muscle_group: MuscleGroup;
  secondary_muscle_groups: Array<MuscleGroup>;
};
// ...
export type MuscleGroup =
  | "Chest"
  | "Back"
  | "Traps"
  | "Shoulders"
  | "Quads"
  | "Hamstrings"
  | "Glutes"
  | "Biceps"
  | "Triceps"
  | "Calves";

/**
 * Input enum mirroring [`MuscleGroup`].
 */
export type MuscleGroupInput =
  | "Chest"
  | "Back"
  | "Traps"
  | "Shoulders"
  | "Quads"
  | "Hamstrings"
  | "Glutes"
  | "Biceps"
  | "Triceps"
  | "Calves";

export type Name = string;

export type PercentOneRepMax = number;

export type Phase = "Accumulation" | "Intensification" | "Deload";
// ...
export type Prescription = {
  microcycle_id: number;
  set_groups: Array<SetGroup>;
};